+++ packages/iov-dpos/src/parse.spec.ts
@@ -1,7 +1,6 @@
 import { ReadonlyDate } from "readonly-date";
 
 import { Nonce } from "@iov/bcp";
-import { Int53 } from "@iov/encoding";
 
 import { Parse } from "./parse";
 
@@ -66,31 +65,28 @@ describe("Parse", () => {
 
   it("can convert time to nonce", () => {
     // zero, positive, negative
-    expect(Parse.timeToNonce(new ReadonlyDate(ReadonlyDate.UTC(1970, 0, 1, 0, 0, 0)))).toEqual(new Int53(
-      0,
-    ) as Nonce);
-    expect(Parse.timeToNonce(new ReadonlyDate(ReadonlyDate.UTC(1999, 2, 3, 4, 5, 6)))).toEqual(new Int53(
-      920433906,
-    ) as Nonce);
-    expect(Parse.timeToNonce(new ReadonlyDate(ReadonlyDate.UTC(1969, 11, 31, 23, 15, 0)))).toEqual(new Int53(
-      -45 * 60,
-    ) as Nonce);
+    expect(Parse.timeToNonce(new ReadonlyDate(ReadonlyDate.UTC(1970, 0, 1, 0, 0, 0)))).toEqual(0 as Nonce);
+    expect(Parse.timeToNonce(new ReadonlyDate(ReadonlyDate.UTC(1999, 2, 3, 4, 5, 6)))).toEqual(
+      920433906 as Nonce,
+    );
+    expect(Parse.timeToNonce(new ReadonlyDate(ReadonlyDate.UTC(1969, 11, 31, 23, 15, 0)))).toEqual((-45 *
+      60) as Nonce);
 
     // beyond year 2038
-    expect(Parse.timeToNonce(new ReadonlyDate(ReadonlyDate.UTC(2040, 0, 1, 0, 0, 0)))).toEqual(new Int53(
-      2208988800,
-    ) as Nonce);
+    expect(Parse.timeToNonce(new ReadonlyDate(ReadonlyDate.UTC(2040, 0, 1, 0, 0, 0)))).toEqual(
+      2208988800 as Nonce,
+    );
 
     // Out of Lisk timestamp range (2016 +/- 70 years). This is not strictly
     // required but shows that we cover the full range and more by storing
     // signed, long UNIX timestamps.
     //
     // python3 -c 'import calendar, datetime; print(calendar.timegm(datetime.datetime(1946, 4, 25, 23, 15, 14, 0).utctimetuple()))'
-    expect(Parse.timeToNonce(new ReadonlyDate(ReadonlyDate.UTC(1946, 3, 25, 23, 15, 14)))).toEqual(new Int53(
-      -747449086,
-    ) as Nonce);
-    expect(Parse.timeToNonce(new ReadonlyDate(ReadonlyDate.UTC(2086, 3, 25, 23, 15, 14)))).toEqual(new Int53(
-      3670614914,
-    ) as Nonce);
+    expect(Parse.timeToNonce(new ReadonlyDate(ReadonlyDate.UTC(1946, 3, 25, 23, 15, 14)))).toEqual(
+      -747449086 as Nonce,
+    );
+    expect(Parse.timeToNonce(new ReadonlyDate(ReadonlyDate.UTC(2086, 3, 25, 23, 15, 14)))).toEqual(
+      3670614914 as Nonce,
+    );
   });
 });