+++ packages/iov-lisk/src/liskconnection.spec.ts
@@ -14,6 +14,7 @@ import {
   isBlockInfoPending,
   isConfirmedTransaction,
   isSendTransaction,
+  Nonce,
   PubkeyQuery,
   PublicKeyBundle,
   PublicKeyBytes,
@@ -225,8 +226,8 @@ describe("LiskConnection", () => {
         const query: AddressQuery = { address: "6472030874529564639L" as Address };
         const nonce = await connection.getNonce(query);
         // nonce is current unix timestamp +/- 300ms
-        expect(nonce.toNumber()).toBeGreaterThanOrEqual(Math.floor(Date.now() / 1000 - 0.3));
-        expect(nonce.toNumber()).toBeLessThanOrEqual(Math.floor(Date.now() / 1000 + 0.3));
+        expect(nonce).toBeGreaterThanOrEqual(Math.floor(Date.now() / 1000 - 0.3));
+        expect(nonce).toBeLessThanOrEqual(Math.floor(Date.now() / 1000 + 0.3));
       }
 
       // by pubkey
@@ -241,8 +242,8 @@ describe("LiskConnection", () => {
         };
         const nonce = await connection.getNonce(query);
         // nonce is current unix timestamp +/- 300ms
-        expect(nonce.toNumber()).toBeGreaterThanOrEqual(Math.floor(Date.now() / 1000 - 0.3));
-        expect(nonce.toNumber()).toBeLessThanOrEqual(Math.floor(Date.now() / 1000 + 0.3));
+        expect(nonce).toBeGreaterThanOrEqual(Math.floor(Date.now() / 1000 - 0.3));
+        expect(nonce).toBeLessThanOrEqual(Math.floor(Date.now() / 1000 + 0.3));
       }
 
       connection.disconnect();
@@ -273,8 +274,8 @@ describe("LiskConnection", () => {
         const nonces = await connection.getNonces(addressQuery, 1);
         expect(nonces.length).toEqual(1);
         // nonce is current unix timestamp +/- 300ms
-        expect(nonces[0].toNumber()).toBeGreaterThanOrEqual(Math.floor(Date.now() / 1000 - 0.3));
-        expect(nonces[0].toNumber()).toBeLessThanOrEqual(Math.floor(Date.now() / 1000 + 0.3));
+        expect(nonces[0]).toBeGreaterThanOrEqual(Math.floor(Date.now() / 1000 - 0.3));
+        expect(nonces[0]).toBeLessThanOrEqual(Math.floor(Date.now() / 1000 + 0.3));
       }
 
       // by address, 2 nonces
@@ -282,9 +283,9 @@ describe("LiskConnection", () => {
         const nonces = await connection.getNonces(addressQuery, 2);
         expect(nonces.length).toEqual(2);
         // last nonce is current unix timestamp +/- 300ms
-        expect(nonces[1].toNumber()).toBeGreaterThanOrEqual(Math.floor(Date.now() / 1000 - 0.3));
-        expect(nonces[1].toNumber()).toBeLessThanOrEqual(Math.floor(Date.now() / 1000 + 0.3));
-        expect(nonces[0].toNumber()).toEqual(nonces[1].toNumber() - 1);
+        expect(nonces[1]).toBeGreaterThanOrEqual(Math.floor(Date.now() / 1000 - 0.3));
+        expect(nonces[1]).toBeLessThanOrEqual(Math.floor(Date.now() / 1000 + 0.3));
+        expect(nonces[0]).toEqual((nonces[1] - 1) as Nonce);
       }
 
       // by address, 3 nonces
@@ -292,10 +293,10 @@ describe("LiskConnection", () => {
         const nonces = await connection.getNonces(addressQuery, 3);
         expect(nonces.length).toEqual(3);
         // last nonce is current unix timestamp +/- 300ms
-        expect(nonces[2].toNumber()).toBeGreaterThanOrEqual(Math.floor(Date.now() / 1000 - 0.3));
-        expect(nonces[2].toNumber()).toBeLessThanOrEqual(Math.floor(Date.now() / 1000 + 0.3));
-        expect(nonces[1].toNumber()).toEqual(nonces[2].toNumber() - 1);
-        expect(nonces[0].toNumber()).toEqual(nonces[1].toNumber() - 1);
+        expect(nonces[2]).toBeGreaterThanOrEqual(Math.floor(Date.now() / 1000 - 0.3));
+        expect(nonces[2]).toBeLessThanOrEqual(Math.floor(Date.now() / 1000 + 0.3));
+        expect(nonces[1]).toEqual((nonces[2] - 1) as Nonce);
+        expect(nonces[0]).toEqual((nonces[1] - 1) as Nonce);
       }
 
       // by pubkey, 0 nonces
@@ -309,8 +310,8 @@ describe("LiskConnection", () => {
         const nonces = await connection.getNonces(pubkeyQuery, 1);
         expect(nonces.length).toEqual(1);
         // nonce is current unix timestamp +/- 300ms
-        expect(nonces[0].toNumber()).toBeGreaterThanOrEqual(Math.floor(Date.now() / 1000 - 0.3));
-        expect(nonces[0].toNumber()).toBeLessThanOrEqual(Math.floor(Date.now() / 1000 + 0.3));
+        expect(nonces[0]).toBeGreaterThanOrEqual(Math.floor(Date.now() / 1000 - 0.3));
+        expect(nonces[0]).toBeLessThanOrEqual(Math.floor(Date.now() / 1000 + 0.3));
       }
 
       // by pubkey, 2 nonces
@@ -318,9 +319,9 @@ describe("LiskConnection", () => {
         const nonces = await connection.getNonces(pubkeyQuery, 2);
         expect(nonces.length).toEqual(2);
         // last nonce is current unix timestamp +/- 300ms
-        expect(nonces[1].toNumber()).toBeGreaterThanOrEqual(Math.floor(Date.now() / 1000 - 0.3));
-        expect(nonces[1].toNumber()).toBeLessThanOrEqual(Math.floor(Date.now() / 1000 + 0.3));
-        expect(nonces[0].toNumber()).toEqual(nonces[1].toNumber() - 1);
+        expect(nonces[1]).toBeGreaterThanOrEqual(Math.floor(Date.now() / 1000 - 0.3));
+        expect(nonces[1]).toBeLessThanOrEqual(Math.floor(Date.now() / 1000 + 0.3));
+        expect(nonces[0]).toEqual((nonces[1] - 1) as Nonce);
       }
 
       // by pubkey, 3 nonces
@@ -328,10 +329,10 @@ describe("LiskConnection", () => {
         const nonces = await connection.getNonces(pubkeyQuery, 3);
         expect(nonces.length).toEqual(3);
         // last nonce is current unix timestamp +/- 300ms
-        expect(nonces[2].toNumber()).toBeGreaterThanOrEqual(Math.floor(Date.now() / 1000 - 0.3));
-        expect(nonces[2].toNumber()).toBeLessThanOrEqual(Math.floor(Date.now() / 1000 + 0.3));
-        expect(nonces[1].toNumber()).toEqual(nonces[2].toNumber() - 1);
-        expect(nonces[0].toNumber()).toEqual(nonces[1].toNumber() - 1);
+        expect(nonces[2]).toBeGreaterThanOrEqual(Math.floor(Date.now() / 1000 - 0.3));
+        expect(nonces[2]).toBeLessThanOrEqual(Math.floor(Date.now() / 1000 + 0.3));
+        expect(nonces[1]).toEqual((nonces[2] - 1) as Nonce);
+        expect(nonces[0]).toEqual((nonces[1] - 1) as Nonce);
       }
 
       connection.disconnect();